requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
jinja2==3.1.2
dropbox==11.36.2
gspread==6.0.0
google-auth==2.27.0
python-pptx==0.6.23
pandas==2.1.4
python-dotenv==1.0.0
setuptools

